Revert "Change `LoginPasswordlessCredentialResponse` index field to int32"
This reverts commit 788e77068102e70f81335fc11338b3b256bc713e.

Revert "Use int32 in tshd check too"

This reverts commit 6a9e6a0c878735a9e2dc821b69e7a7164e1ca43e.

+++ lib/teleterm/clusters/cluster_auth.go
@@ -403,7 +403,7 @@ func (p *pwdlessLoginPrompt) PromptCredential(deviceCreds []*wancli.CredentialIn
 
 	// Test for out of range index values.
 	selectedIndex := credRes.GetIndex()
-	if selectedIndex < 0 || selectedIndex > int32(len(creds))-1 {
+	if selectedIndex < 0 || selectedIndex > int64(len(creds))-1 {
 		return nil, trace.BadParameter("invalid login name")
 	}
